PYIC-7977 Rename ticf cri package
